+++ CHANGELOG.md
@@ -6,6 +6,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [v5.0-beta.10_schema-v4.0-beta.1] - 2021-01-08
+
+* Add `Purpose` arg to `loggedXXX` methods.
+
+* Add example application
+
+* Improve README
+
+
 ## [v5.0-beta.9_schema-v4.0-beta.1] - 2021-01-07
 
 * Add missing handling of unsuccessful logged work outcome.
@@ -152,7 +161,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [v3.1.0] - 2017-04-05
 Intial open source release
 
-[Unreleased]: https://github.com/gchq/event-logging/compare/v5.0-beta.9_schema-v4.0-beta.1...HEAD
+[Unreleased]: https://github.com/gchq/event-logging/compare/v5.0-beta.10_schema-v4.0-beta.1...HEAD
+[v5.0-beta.10_schema-v4.0-beta.1]: https://github.com/gchq/event-logging/compare/v5.0-beta.9_schema-v4.0-beta.1...v5.0-beta.10_schema-v4.0-beta.1
 [v5.0-beta.9_schema-v4.0-beta.1]: https://github.com/gchq/event-logging/compare/v5.0-beta.8_schema-v4.0-beta.1...v5.0-beta.9_schema-v4.0-beta.1
 [v5.0-beta.8_schema-v4.0-beta.1]: https://github.com/gchq/event-logging/compare/v5.0-beta.7_schema-v4.0-beta.1...v5.0-beta.8_schema-v4.0-beta.1
 [v5.0-beta.7_schema-v4.0-beta.1]: https://github.com/gchq/event-logging/compare/v5.0-beta.6_schema-v4.0-beta.1...v5.0-beta.7_schema-v4.0-beta.1